Updated db/index.js to include pool for the connection.js, include a class of DB for async query, and include a findAllEmployees method for the query/Updated index.js to include the required packages for inquirer as an object of prompt, include a logo variable for requiring the asciiart-logo so I can add a splashscreen for the data tables, an init function to initialize the splash screen, and added a loadMainPrompts function with a prompt question/Updated schema.sql to include a working database sequence that will actually load up the datatable when I \i db/schema.sql in the CLI however I cannot get the seeds.sql to insert into the tables will fix.
Updated db/index.js to include pool for the connection.js, include a class of DB for async query, and include a findAllEmployees method for the query. Updated index.js to include the required packages for inquirer as an object of prompt, include a logo variable for requiring the asciiart-logo so I can add a splashscreen for the data tables, an init function to initialize the splash screen, and added a loadMainPrompts function with a prompt question. Updated schema.sql to include a working database sequence that will actually load up the datatable when I \i db/schema.sql in the CLI however I cannot get the seeds.sql to insert into the tables will fix.