Update schema.sql to remove a comma on line 24, add comments to provide clarity, changed the name of the departments_db to employees_db, and moved the role_id foreign key so that it will be above the manager key reference added as well/Updated connection.js to include client then use const client = new Client to make a new client for the host then I used the connect method to connect the client, added a client.query method to select all the data from the departments table and then it handles the rows by console logging them if there is no error, else console log error message, then end client query, added an export for client/Updated seeds.sql to fix not adding in the values however it has not been fixed yet will fix/Updated db/index.js to fix up line 19 so it properly selects the employee data, and then joins it into the departments table however I'm not sure that this works yet because I have not completed the prompts will fix.
Update schema.sql to remove a comma on line 24, add comments to provide clarity, changed the name of the departments_db to employees_db, and moved the role_id foreign key so that it will be above the manager key reference added as well. Updated connection.js to include client then use const client = new Client to make a new client for the host then I used the connect method to connect the client, added a client.query method to select all the data from the departments table and then it handles the rows by console logging them if there is no error, else console log error message, then end client query, added an export for client. Updated seeds.sql to fix not adding in the values however it has not been fixed yet will fix. Updated db/index.js to fix up line 19 so it properly selects the employee data, and then joins it into the departments table however I'm not sure that this works yet because I have not completed the prompts will fix.