Updated index.js to pseudo code next steps for the prompts, and how to tie in the functions that will return the field and records for the tables, and added in a switch statement that takes in the userdata from the prompted questions, added a function to find all employees/Updated db/index.js to pseudo code in what to do with the interpolation for adding stuff to the tables during the prompts, and then also add a comment to make sure the database is exported to the index.js so the functions for the prompts can be used.
Updated index.js to pseudo code next steps for the prompts, and how to tie in the functions that will return the field and records for the tables, and added in a switch statement that takes in the userdata from the prompted questions, added a function to find all employees. Updated db/index.js to pseudo code in what to do with the interpolation for adding stuff to the tables during the prompts, and then also add a comment to make sure the database is exported to the index.js so the functions for the prompts can be used.