Updated db/connection.js to include parents on the client.end so that when called it will close the database connection/Updated db/index.js to include functions for viewing the departments, roles, and employees using an async function that will return to the prompt data that will be viewable and relevant to the question however I have not used the exported file in the index.js to add the functions from db/index.js to the main prompt will fix.
Updated db/connection.js to include parents on the client.end so that when called it will close the database connection. Updated db/index.js to include functions for viewing the departments, roles, and employees using an async function that will return to the prompt data that will be viewable and relevant to the question however I have not used the exported file in the index.js to add the functions from db/index.js to the main prompt will fix.