Updated db/connection.js to comment out, then removed code from 10-21 because it was not needed, and changed the Client object variable to Pool so I could have a new Pool for each client/Updated db/index.js to add comments on line 8 and 13, added the findAllDepartments function to select the id and name from department, added the findAllRoles function to select the role title as job title grab the role id the department name the role salary from the role table then join the department on the role.department id as department id, added the findAllEmployees function to get employee ids, first names, last names, job titles, departments, salaries, and managers however I cant get the managers to go into the correct row will fix, added the create department so I can insert into department a new department name value using interpolation, added the createRole function to insert a new role into the role table with a new department name, role title and role salary using template literal interpolation, added the create employee function however the role and manager are not populating correctly will fix, added the update employee role by selecting an existing employee and updating their role title however the employee selection will be implemented will fix, and added comments for clarity/Updated index.js to implement each function from the database, changed line 10 to match the applications brand, added a comment on line 4 to explain what requiring the database does, added the prompts to the loadMainPrompt function using inquirer, added a switch case to handle the database query functions messaging for console with handling for the function to relate to the databases query functions in the database, and then added each function below to handle adding and viewing the data necessary for the table, and added comments for clarity/Updated the seeds.sql in the data base so that the values that populate the schema function correctly by changing the spelling errors, and added a comment for clarity/Updated the schema.sql in the database to add comments for clarity, and update the spelling errors between the query and the tables created the schema was plural and now it is not.
Updated db/connection.js to comment out, then removed code from 10-21 because it was not needed, and changed the Client object variable to Pool so I could have a new Pool for each client. Updated db/index.js to add comments on line 8 and 13, added the findAllDepartments function to select the id and name from department, added the findAllRoles function to select the role title as job title grab the role id the department name the role salary from the role table then join the department on the role.department id as department id, added the findAllEmployees function to get employee ids, first names, last names, job titles, departments, salaries, and managers however I cant get the managers to go into the correct row will fix, added the create department so I can insert into department a new department name value using interpolation, added the createRole function to insert a new role into the role table with a new department name, role title and role salary using template literal interpolation, added the create employee function however the role and manager are not populating correctly will fix, added the update employee role by selecting an existing employee and updating their role title however the employee selection will be implemented will fix, and added comments for clarity. Updated index.js to implement each function from the database, changed line 10 to match the applications brand, added a comment on line 4 to explain what requiring the database does, added the prompts to the loadMainPrompt function using inquirer, added a switch case to handle the database query functions messaging for console with handling for the function to relate to the databases query functions in the database, and then added each function below to handle adding and viewing the data necessary for the table, and added comments for clarity. Updated the seeds.sql in the data base so that the values that populate the schema function correctly by changing the spelling errors, and added a comment for clarity. Updated the schema.sql in the database to add comments for clarity, and update the spelling errors between the query and the tables created the schema was plural and now it is not.