Refactor:(docs): Corrected the NULL values to associate with correct manager in db/seeds.sql/Fix:(Selector): Fixed the way data is selected through the functions in db/index.js/Fix:(Prompts): Fixed the prompts messaging and error handling for creating a department, creating a role, creating an employee, and updating an employee.
Refactor:(docs): Corrected the NULL values to associate with correct manager in db/seeds.sql.

Fix:(Selector): Fixed the way data is selected through the functions in db/index.js.

Fix:(Prompts): Fixed the prompts messaging and error handling for creating a department, creating a role, creating an employee, and updating an employee.